+++ activesupport/lib/active_support/core_ext/array.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require "active_support/core_ext/array/wrap"
-require "active_support/core_ext/array/access"
-require "active_support/core_ext/array/conversions"
-require "active_support/core_ext/array/extract"
-require "active_support/core_ext/array/extract_options"
-require "active_support/core_ext/array/grouping"
-require "active_support/core_ext/array/inquiry"
+require_relative "array/wrap"
+require_relative "array/access"
+require_relative "array/conversions"
+require_relative "array/extract"
+require_relative "array/extract_options"
+require_relative "array/grouping"
+require_relative "array/inquiry"