+++ railties/lib/rails/gem_version.rb
@@ -10,7 +10,7 @@ module Rails
     MAJOR = 8
     MINOR = 1
     TINY  = 0
-    PRE   = "beta1"
+    PRE   = "rc1"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end